+++ R/addin.R
@@ -210,6 +210,15 @@ read_utf8 <- function(x) {
   paste(readLines(x, encoding = "UTF-8"), collapse = "\n")
 }
 
+# c.f. https://kevinushey.github.io/blog/2018/02/21/string-encoding-and-r/
+write_utf8 <- function(x, f) {
+  # Ensure the text is encoded as UTF-8
+  x <- enc2utf8(x)
+
+  con <- withr::local_connection(file(f, open = "w+", encoding = "native.enc"))
+  writeLines(x, con = con, useBytes = TRUE)
+}
+
 extract_image_paths <- function(html_text) {
   html_doc <- xml2::read_html(html_text)
   img_nodes <- xml2::xml_find_all(html_doc, ".//img")